Making the code more mac-friendly, due to the way DWORD is defined in ftdi headers.

+++ xid_device_driver/interface_xid_con.h
@@ -1,6 +1,8 @@
 #ifndef INTERFACE_XID_CON_T_H
 #define INTERFACE_XID_CON_T_H
 
+#include "ftd2xx.h"
+
 namespace cedrus
 {
     class interface_xid_con
@@ -9,7 +11,7 @@ namespace cedrus
         virtual ~interface_xid_con()
         {}
 
-        virtual bool read( unsigned char *in_buffer, int bytes_to_read, unsigned long *bytes_read) = 0;
+        virtual bool read( unsigned char *in_buffer, int bytes_to_read, DWORD *bytes_read) = 0;
 
     };
 } // namespace cedrus
